Fix BOOST 1.60 warns

+++ tests/tests/atomicswap_check/atomicswap_helper_check.cpp
@@ -72,17 +72,17 @@ SCORUM_TEST_CASE(secret_generator)
 {
     std::string entropy(SCORUM_ATOMICSWAP_SECRET_MAX_LENGTH * 10, ' ');
 
-    BOOST_CHECK_LE(get_secret_hex(entropy).size(), SCORUM_ATOMICSWAP_SECRET_MAX_LENGTH);
+    BOOST_CHECK_LE(get_secret_hex(entropy).size(), (size_t)SCORUM_ATOMICSWAP_SECRET_MAX_LENGTH);
 
     entropy = std::string(SCORUM_ATOMICSWAP_SECRET_MAX_LENGTH / 10, ' ');
 
-    BOOST_CHECK_LE(get_secret_hex(entropy).size(), SCORUM_ATOMICSWAP_SECRET_MAX_LENGTH);
+    BOOST_CHECK_LE(get_secret_hex(entropy).size(), (size_t)SCORUM_ATOMICSWAP_SECRET_MAX_LENGTH);
 
     fc::sha512 hash;
 
     BOOST_REQUIRE(hash.data_size() >= 32);
 
-    BOOST_CHECK_EQUAL(get_secret_hex(entropy, 32).size(), 32 * 2);
+    BOOST_CHECK_EQUAL(get_secret_hex(entropy, 32).size(), (size_t)32 * 2);
 
     BOOST_CHECK_EQUAL(get_secret_hex(entropy, hash.data_size()).size(), hash.data_size() * 2);
